crawler/music_crawler/requirements.txt
scrapy==2.12.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
celery[redis]==5.5.1
redis==5.2.1
qdrant-client==1.13.3